+++ src/__mocks__/tone.js
@@ -234,6 +234,38 @@ class AutoPanner {
 }
 
 // ----------------------------------------------------------------------------
+// Tone.EQ3
+// ----------------------------------------------------------------------------
+
+export const mockEQ3Constructor = jest.fn();
+
+class EQ3 {
+  constructor(low, mid, high) {
+    mockEQ3Constructor(low, mid, high);
+
+    this.low = {
+      value: low,
+    };
+
+    this.mid = {
+      value: mid,
+    };
+
+    this.high = {
+      value: high,
+    };
+
+    this.lowFrequency = {
+      value: 400,
+    };
+
+    this.highFrequency = {
+      value: 2500,
+    };
+  }
+}
+
+// ----------------------------------------------------------------------------
 // Tone.Sequence
 // ----------------------------------------------------------------------------
 
@@ -268,6 +300,7 @@ const MockTone = {
   Sampler,
   AutoFilter,
   AutoPanner,
+  EQ3,
   Sequence,
 };